requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
psycopg2==2.8.1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
